backend/requirements.txt
# ScheduleLink Backend Dependencies
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
pydantic[email]==2.6.1
pydantic-settings==2.1.0
stripe==8.4.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.118.0
aiosmtplib==3.0.1
python-dotenv==1.0.1
httpx==0.26.0
apscheduler==3.10.4
